Refactor dependency middleware.
